Repair tests.

+++ SingleSource/UnitTests/2002-05-19-DivTest.c
@@ -1,3 +1,4 @@
+// Modified by LLVM-MOS.
 
 #ifndef __STDC_LIMIT_MACROS
 #define __STDC_LIMIT_MACROS 1
@@ -7,19 +8,19 @@
 extern int printf(const char *, ...);
 
 void testL(int64_t Arg) {
-  printf("%d\n", (int) (Arg / ((int64_t)1 << 4)));
-  printf("%d\n", (int) (Arg / ((int64_t)1 << 46)));
+  printf("%ld\n", (long) (Arg / ((int64_t)1 << 4)));
+  printf("%ld\n", (long) (Arg / ((int64_t)1 << 46)));
 }
 
-void test(int Arg) {
-  printf("%d\n", Arg / (1 << 0));
-  printf("%d\n", Arg / (1 << 4));
-  printf("%d\n", Arg / (1 << 18));
-  printf("%d\n", Arg / (1 << 30));
+void test(long Arg) {
+  printf("%ld\n", Arg / (1L << 0));
+  printf("%ld\n", Arg / (1L << 4));
+  printf("%ld\n", Arg / (1L << 18));
+  printf("%ld\n", Arg / (1L << 30));
 }
 
 int main() {
-  int B20 = - (1 << 20);
+  long B20 = - (1L << 20);
   int64_t B53 = - ((int64_t)1 << 53);
 
   test(B20 + 32);
